trading-robot/requirements.txt
autopep8==2.0.2
cachetools==5.3.0
certifi==2023.5.7
charset-normalizer==3.1.0
codecov==2.1.13
contourpy==1.0.7
coverage==7.2.5
cycler==0.11.0
deprecation==2.1.0
filelock==3.12.2
fonttools==4.39.4
grpcio==1.54.2
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.0
psycopg2==2.9.6
pycodestyle==2.10.0
pydantic==1.10.7
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.30.0
six==1.16.0
sympy==1.12
tinkoff==0.1.1
tinkoff-investments==0.2.0b57
tomli==2.0.1
torch==2.0.1
typing_extensions==4.5.0
tzdata==2023.3
urllib3==2.0.2
